Removed old CfgPatches
