+++ BNA_KC_Weapons/Config.cpp
@@ -1,23 +1,16 @@
-#include "CfgPatches.hpp"
 #include "CfgFunctions.hpp"
 
 
-class Mode_SemiAuto;
-class Mode_FullAuto: Mode_SemiAuto {};
-
-class SlotInfo {};
-class CowsSlot: SlotInfo {};
-class CowsSlot_Rail: CowsSlot {};
-class PointerSlot: SlotInfo {};
-class PointerSlot_Rail: PointerSlot {};
-class UnderBarrelSlot: SlotInfo {};
-class UnderBarrelSlot_Rail: UnderBarrelSlot {};
-
-
-class CfgMagazines
+class CfgPatches
 {
-};
-
-class CfgAmmo
-{
+    class BNA_KC_Weapons
+    {
+        author = "SweMonkey and DartRuffian";
+        requiredVersion = 0.1;
+        requiredAddons[] = {};
+        units[] = {};
+        weapons[] = {};
+        magazines[] = {};
+        ammo[] = {};
+    };
 };
